+++ src/ctest.ts
@@ -1259,11 +1259,14 @@ export class CTestDriver implements vscode.Disposable {
         // Commands can't be used to replace array (i.e., args); and both test program and test args requires folder and
         // test name as parameters, which means one lauch config for each test. So replacing them here is a better way.
         chosenConfig.config = this.replaceAllInObject<vscode.DebugConfiguration>(chosenConfig.config, '${cmake.testProgram}', this.testProgram(testName));
+        chosenConfig.config = this.replaceAllInObject<vscode.DebugConfiguration>(chosenConfig.config, '${command:cmake.testProgram}', this.testProgram(testName));
         chosenConfig.config = this.replaceAllInObject<vscode.DebugConfiguration>(chosenConfig.config, '${cmake.testWorkingDirectory}', this.testWorkingDirectory(testName));
+        chosenConfig.config = this.replaceAllInObject<vscode.DebugConfiguration>(chosenConfig.config, '${command:cmake.testWorkingDirectory}', this.testWorkingDirectory(testName));
 
         // Replace cmake.testArgs wrapped in quotes, like `"${command:cmake.testArgs}"`, without any spaces in between,
         // since we need to repalce the quotes as well.
         chosenConfig.config = this.replaceArrayItems(chosenConfig.config, '${cmake.testArgs}', this.testArgs(testName)) as vscode.DebugConfiguration;
+        chosenConfig.config = this.replaceArrayItems(chosenConfig.config, '${command:cmake.testArgs}', this.testArgs(testName)) as vscode.DebugConfiguration;
 
         // Identify the session we started
         chosenConfig.config[magicKey] = magicValue;
@@ -1307,7 +1310,7 @@ export class CTestDriver implements vscode.Disposable {
         }
     }
 
-    private testProgram(testName: string): string {
+    public testProgram(testName: string): string {
         if (this.tests) {
             for (const test of this.tests.tests) {
                 if (test.name === testName) {
@@ -1324,7 +1327,7 @@ export class CTestDriver implements vscode.Disposable {
         return '';
     }
 
-    private testWorkingDirectory(testName: string): string {
+    public testWorkingDirectory(testName: string): string {
         const property = this.tests?.tests
             .find(test => test.name === testName)?.properties
             .find(prop => prop.name === 'WORKING_DIRECTORY');
@@ -1335,7 +1338,7 @@ export class CTestDriver implements vscode.Disposable {
         return '';
     }
 
-    private testArgs(testName: string): string[] {
+    public testArgs(testName: string): string[] {
         if (this.tests) {
             for (const test of this.tests.tests) {
                 if (test.name === testName) {